split mocked and non mocked tests into two separate modules
